[startup] correctly set the classpath in dev build for content modules when module-based loader is used and modules are compiled by Bazel (IJPL-205494)
When modules are compiled by Bazel, the module containing plugin.xml is packed to a JAR file in out/dev-run/jar-cache, so 'pluginDir' is set to out/dev-run directory. The JAR for the content module won't be found there, so `resolveCustomModuleClassesRoots` is now used to take the proper data from the runtime module repository.

GitOrigin-RevId: ec718a88bebed43e0c53b88c58d3b265431e46b6