[intellij project configuration] enable 'Suspicious package-private access' in test sources as well (IJPL-205216)
Tests which access package-private members from production code of the same module will fail to compile with Bazel. So this inspection is useful to see such errors in the editor. Also, if a test uses a protected member from a production code of the same module not inside a subclass, it fails with IllegalAccessError if tests are loaded by a different classloader (see IJPL-116621), so it's better to highlight such cases as well.

All reports in the intellij monorepo except intellij.android.* modules are fixed.

GitOrigin-RevId: 2920ecef9bf5a3985d89e836f152de86a0d097c9