[maven][eel][jps][IDEA-375248] fix stale JPS compilation cache on a remote target
The error was caused by reusing the old cache for compilation. As a result, the file with the compiler settings did not match the current state of the project. This led to the module being compiled with the version of the target Java bytecode that was used for the first compilation.

GitOrigin-RevId: 3a5e40192f664ca842750b2d276b7b4b44d67eaa