[terminal] IJPL-195475 Fix cursor blinking
There was annoying blinking because of type-ahead and the implementation details of the `TerminalCursorPainter`.
Before, we updated the cursor painter also on output model content change. But it this moment cursor positon might be incorrect, so it created a blinking effect.
Now we can remove this hack, since we receive all cursor position change updates thanks to `TerminalOffset`.

GitOrigin-RevId: 82396ca1f50fe8d2b2f7644c1d40da81c4fc9178