IJPL-198457 Search Everywhere: Item with UUID is not found in the list
The bug was happening, because we let RecentFiles results fast pass the throttling. That caused the order change: class MyClass could be generated before the recent file MyClass.java on backend and reach equality checker earlier, but on the frontend the order changes, and uuidToReplace from MyClass.java element is not yet in the list.

Now we save uuids to the pending replacements and ignore the arrived later elements if necessary.

GitOrigin-RevId: 08987059a18894a3a1eb790b1b2ba8e8083cc1c5