[build script tests] diagnostics: publish log files of IDEs started from build scripts in tests as build artifacts
Before, if an IDE process started from build scripts failed, its log file wasn't published if build scripts are started from tests. Now, these logs are stored in the build logs directory, and the whole content of that directory is published as build artifacts in BuildScriptTestUtilsKt.doRunTestBuild if the build fails.

GitOrigin-RevId: dd26d9c14faa5eec81c7be0f0cb1fcb05d2f4132