[junit5] Support reusing abstract static fixtures across tests
The static test fixture defined in the abstract class was initialized once but tear down after the first inheritor of the abstract class finished executing. That meant that all the following inheritors would receive the feature in the disposed state. Resetting the state of the feature after each tear down allows reinitializing and reusing it.

GitOrigin-RevId: eee2a5f1bf8ffc18bed0817999e45b76359e5861