IJPL-167857 Add a separate action for double click
When “Select Word at Caret” is invoked by double-click, `CaretImpl` passes the current caret to the action handler, so only the word under that caret is selected.
However, the action can also be executed via the regular shortcut (“Extend Selection”), in which case the selection is extended for each active caret.

In Remote Dev, we can track and redirect action execution but not double-clicks. Therefore, when multiple carets are active, the selection is extended for all of them.

To fix that, we can introduce a separate action which is invoked only on double click. In this case the action sysyem in RD can re-direct this action execution to backend without any problems.

GitOrigin-RevId: 53268a1c658a4eb78080268aa1ba0d1310e72f79