[terminal] IJPL-205931 Fix recursive initialization of the terminal output document
It was not a good idea to get the current document state from the `FileViewProvider`. Since the document might not be present there and it will decide to create one. And will try to use the PSI content as the initial state of the document. And the PSI file will try to get the document, so there is the loop.

Let's not depend on the document state in the PSI file directly. Instead, store the `charsSequence` reference to the recent document snapshot and update it in the file on output model changes.

GitOrigin-RevId: 7f25691c7425cc4efc67ea50aed8aa8ed4482b98