[module-based loader] fix loading plugins when running build tests from compiled archives (IJI-2806)
If the module-based loader is used to start the fronted process not from a distribution, but from class-files packed in per-module archives, it may happen that there is only one JAR file included in PluginModuleGroup, but some content module depending on the backend is present in plugin.xml. In that case, ModuleBasedPluginXmlPathResolver should be used as well to ensure that loading plugin.xml won't fail at runtime because of missing XML descriptor for such a content module.

This fixes IdeaUltimateAdditionalBuildTest.jetBrainsClientForIdeaBuildSuccessfully when running with 'USE_ARCHIVED_COMPILED_CLASSES' set to 'true'.

GitOrigin-RevId: 43671471c97e28f8b0948a8d907e9884d620305b