[build scripts] change JAR names in the IDE_HOME/lib directory for IDEs with embedded client (IJPL-170362)
Before, '-client' suffix was used to JARs in IDE_HOME/lib which should be included in the classpath of the embedded frontend (JetBrains Client) process, and no suffix was specified for JARs which should be included in the  IDE running in monolith or backend mode. Now for consistency with naming schema name without suffix is used for JARs added in the classpath of all processes, and '-backend' suffix is used for JARs used in backend and monolith mode only.

So, now 'app.jar', 'lib.jar' and 'product.jar' are added to the classpath of all processes, and 'app-backend.jar', 'lib-backend.jar', 'product-backend.jar' are added to the classpath for IDE running in backend and monolith mode only.

GitOrigin-RevId: 5423d307cc8573c5a865e173bb394d8aef743e6b