[Gradle][VersionCatalog][WSM] IDEA-370426 get version catalog locations from GradleVersionCatalogEntity
Getting a version catalog location is used for navigation to a version catalog from a build script, for finding usages etc.

Previously, `GradleDslVersionCatalogHandler` was used for getting version catalog locations. It relied on parsing of `settings.gradle(.kts)` files, implemented in `com.android.tools.idea.gradle.dsl`. This approach has some lacks described in comments for IDEA-370426.

In addition to the old approach, this commit implements getting version catalog locations from GradleVersionCatalogEntity, created from data received from Gradle at sync.

However, the old approach is still used in case if GradleVersionCatalogEntity is not found. For example, when the sync was not done yet.

Classes, for which getting version catalog locations was changed:
1) org.jetbrains.plugins.gradle.service.toml.GradleVersionCatalogTomlAwareGotoDeclarationHandler
Is tested by:
* org.jetbrains.plugins.gradle.dsl.versionCatalogs.GradleVersionCatalogsResolveTest
* org.jetbrains.plugins.gradle.dsl.versionCatalogs.GradleVersionCatalogsFindUsagesTest (this test still relies on the old approach with parsing settings files, because it does not trigger the sync)
* org.jetbrains.plugins.gradle.dsl.versionCatalogs.GradleVersionCatalogsRenameTest (the same, the test does not trigger the sync, so it relies on the settings file parsing)

2) org.jetbrains.kotlin.idea.gradle.versionCatalog.toml.KotlinGradleTomlVersionCatalogGotoDeclarationHandler
Is tested by:
* org.jetbrains.kotlin.idea.k2.codeInsight.gradle.versionCatalog.KotlinGradleVersionCatalogNavigationTest
* org.jetbrains.kotlin.idea.k2.codeInsight.gradle.versionCatalog.KotlinGradleVersionCatalogRenameTest
* org.jetbrains.kotlin.idea.k2.codeInsight.gradle.KotlinGradleFindUsagesTest

3) org.jetbrains.plugins.gradle.service.resolve.GradleProjectExtensionContributor
Is tested by:
* org.jetbrains.plugins.gradle.dsl.versionCatalogs.GradleVersionCatalogsCompletionTest

GitOrigin-RevId: 4649d1a76237c10f8213004d5369c75a034f4517