PY-81439 Add a regression test for the problem
The problem itself was fixed together with PY-82979
when we started recognizing Literal types in the heuristic
for implicit type aliases.

The problem didn't reproduce with Literal types containing
only one variant, e.g. `Literal[1]`, because in this case
`PyTypingTypeProvider.getType` returned a PyLiteralType,
as opposed to a union of two literal types for `Literal[1, 2]`,
and the fallback logic in
`PyTypeHintsInspection.Companion.referenceIsCorrectTypeHint`
considered such a type hint valid
(confusing, to be revised separately).

GitOrigin-RevId: a7da296099652b02deaa0a2e4a2e2fd85de94f87