[remdev PV] IJPL-166441 Add save/restore path logging
1. In TreeState, add some higher level debug logging
corresponding to the start of the expand / load operation
and to locating specific paths. Now we'll be able to tell
what went wrong if some paths were not found.

2. On the remdev side, add visitor logging.
Log when a visitor is created and log the paths it finds.
This will help us to understand what exactly
something like "Refresh 12" means, as it will now log
the path corresponding to id=12.

GitOrigin-RevId: 940697e51895c651bc5c7f91f1dbd84152124ccd