[java-psi] Disable nullity processing during upper bound capturing
We rely on PsiCapturedWildcardType identity when creating recursive types (the upper bound of a PsiCapturedWildcardType may refer to the same type). However, when we are under NullMarked annotation, the PsiCapturedWildcardType instance could be recreated, which prevents us from creating a proper PsiType object.
Fixes IDEA-378142 Regress issue with resolve type of returned generics value

GitOrigin-RevId: 4fb3c306e302715ab8de93f606397cd2a3dcb864