IJPL-206785 IJent: set different log levels in the server app
This change not only fixes the mentioned issue, but also it's expected that the performance of IJent increases when debug/trace logs are enabled. Before this commit the IJent server app was sending all debug/trace logs to the IDE, even those log messages that the IDE will certainly filter out. These logs used to pollute the transport channel in WSL and Docker.

In the server app, `#` in log categories was replaced with `-` because `#` is not accepted by `EnvFilter::builder().parse()`.

This commit also removes `trace2!` and `LogLevel::ALL` for two reasons:
* It was difficult to support in the new paradigm.
* The new mechanism gives more control over log filtering. There's no need in `trace2` actually.

GitOrigin-RevId: c04fb11bb2092441eebd52771553ff25fd483e8c