Cherry pick settings-sync commits
This cherry-picks 5 settings-sync commits from the github query:
https://github.com/search?q=repo%3AJetBrains%2Fintellij-community+IJPL-199443&type=commits

We consolidate this into a single commit for easier maintenace.

There were several conflicts when attemping to cherry-pick so I did by
best to preserve the functionality (since there
is quite a bit of divergence from 2025.2.2 and jetbrains master branch
as of two weeks ago), but I think this migrates the wanted behavior.

Additionally, ab49eec8f9 caused issues since LicenseManager is not
available in 2025.2.2 so this was modified and ed63e19479 was dropped.

Change-Id: I14a30c5d051e6f5cf7dbb3d61a624679ddcad3f2