[todo] IJPL-208085 Implement simple usage of listing TODO items through RPC
`TodoFileNode.getChildrenImpl` is a method that retrieves all the TODO values for a certain file (this method is called even in other scopes, just for each file in the scope that has TODO items). It's implementation has been modified to use two different approaches, one in split mode and the other one in monolith mode. `TodoRemoteClient` is used to link the remote case to the RPC implementation.
As frontend needs a PSI-free representation of the tree node, the minimal `TodoRemoteItemNode` is created to have the prebuilt navigation properties in the `Value` object and presentation chunks. The class is still in progress and will be extended.

GitOrigin-RevId: d541fdddf0a6d36f900e02d0a8cf068b1eeca14d