[python PSI] IJPL-206569 Make Python's MyInstanceElement not physical
This is needed, first of all, to make
DjangoAdminFieldReferenceCompletionTest
pass. It starts to fail when deferring icons in ElementBase.getIcon,
because it's impossible to create smart pointers for this element,
the code in SmartPsiElementPointerImpl.createElementInfo
fails, and the comment there tells that isPhysical=false
is the way to solve it.

It's old code, but DjangoAdminFieldReferenceCompletionTest
is new, and it was introduced after ElementBase.getIcon
stopped deferring icons on a BGT. Now that we want
that deferring back, it seems logical to fix this on the PSI side.

GitOrigin-RevId: 023f1849f0b636f82a4200475e94c498ebf1fd41