[Gradle|Sync] fix: specify module SDK also based on IdeaProject.getSdkName
The module SDK should be specified for the Gradle Java project, if there is one. A null module SDK means that something was wrong in Gradle sync. It is the SDK data service's responsibility to decide whether to reuse the project SDK as the module SDK. SDK data nodes should have an effective module SDK.

### Issues
  * IDEA-367680 The module SDK isn't always specified

GitOrigin-RevId: 4d069d704979a4b0f4f87ca5265ecb88f218d281