[compilation charts] cleanup: remove unused plugin alias from a content module (IJPL-207059)
Since other plugins aren't supposed to declare dependencies on this content module, it shouldn't define a plugin. This is because even after we make content modules private by default (see IJPL-207059), a module defining a plugin alias inside can still be referenced via that alias.

There are no usages of this alias in the monorepo and plugins from the Marketplace, so they can be safely removed.

GitOrigin-RevId: 10e83f9c77899561d5457b7a6bfd4fb9bdb3ba8a