[platform, core] IJPL-206569 Don't defer PSI icons in tests
For some reason, a couple of tests in
CustomScopePaneTest
failed in a really strange way: the top directory
of the project stopped being recognized by the project
index as a part of the project content.

It's not clear why that happened, but it seems
like a nice idea to disable icon deferring in tests,
because unlike in production, test may never actually
try to paint the returned icon, and therefore
the actual icon computation won't be tested.

GitOrigin-RevId: 11f91e5395ce47afc2356e562bfd214e08f8c8e2