[platform UI] IJPL-195472 Ditch the ide.tree.bulk.expand.tree.state key
This was an experiment to see what would happen
if we try to wait until the entire tree has loaded.
In practice, it didn't work well as it often required
half a minute. It was replaced by the presentation caching.

Before fixing TreeState bugs, it's nice to get rid of code
we don't need.

GitOrigin-RevId: 8031b7ad1efc9f775210bba4e3f7e8890f9c7443