[platform UI] IJPL-195472 Introduce TreeChildrenPreloader
The idea is to monitor the visible part of the tree
and try to eagerly load one level of children.

It already works fairly well except it doesn't cover
the auto-expand case, when the actually expanded
node is several levels down. Let's wait and see
if it's really an issue.

GitOrigin-RevId: e09fa7babd85644c59fedb02392ec4851048fcf6