[platform UI] IJPL-195472 Fix cached presentation when the root is already loaded
For some models, the root loads very quickly because
it's a synthetic node that is returned immediately.
It's a problem because the promise may get resolved so fast
that it's loaded at the moment the model is set to the tree.
When we then set the cached presentation, it's essentially
ignored because it can't find a match for the root
because it missed the "root loaded" event.

Fix by reporting the root directly to the presentation
when the presentation is set. We could do the same for
all the nodes loaded so far, but it doesn't seem to be an issue.

GitOrigin-RevId: cb29263e0b14d52fb940fe8daa062f7c1dac851d