[platform UI] IJPL-195472 Fix children loading in TreeState
In the case when a cached presentation is applied,
TreeState tries to load the real nodes before clearing
the presentation and reporting "done."
But in the case the model supports async children loading,
the children of the lowest-level expanded nodes can be
not loaded yet because they're not visited,
only the expanded ones are.

Fix by also visiting children. This required a significant
update of the algorithm. Now it's based on a stack
of potential matches, and we decide on the next visitor
action based on what was matched at the current level.

GitOrigin-RevId: 14d64f8bd6cad7fd13be258f8dd3c7f6ce2c7b7e