[JEWEL-1001] Fix excessive line height with IDE zoom
This fixes the line calculation for text in the IDE when the IDE zoom
feature is enabled. We were not measuring the line height correctly in
that case, as it turns out that for LaF fonts we actually need to take
the IDE zoom into account and divide all metrics by it to get the right
result.

Amusingly, this is not the case for editor scheme fonts — those are not
pre-scaled like LaF fonts are, so special accommodations are made for
those.