[run dashboard & split mode] IJPL-184406 Reimplement customizations API
Because of external usages and the specifics of the api it has to remain on the backend side (it touches process handler). To make it work, a customization builder is introduced: it gets passed to registered extensions, accumulates customizations, sends it through rpc and applies to the corresponding frontend UI component.

GitOrigin-RevId: 1d0f1bdc954c4b2837e17fb472aa665f0a36e03c