[execution & split mode] IJPL-184406 Correctly implement content to reuse logic
The old code requires working with run configurations, but it effectively expects only the correct type id and name from it. Hence we create a fake configuration with correct identifying data inside to preserve as much legacy code working as possible, yet making it on the frontend side where the check now gets called.

GitOrigin-RevId: aeb2f5fe2ff2ef6413b26328ec8ef4fcceed58e8