[execution & split mode] IJPL-184406 Add RunContentDescriptorId to the debug session RPC api to enable debugger in split services view
- The debugger is split the way that the real content is on the frontend, the real descriptor is on the backend. But the old code wanted both to be available. The descriptor is primarilly required on the frontend to get its configuration and check whether the corresponding content must be shown in services view or in a dedicated toolwindow. Since this information is lost in the frontend debugger implementation, we add the RunContentDescriptorId to the session api.

When the debugger tries to create a real frontend content, it passes the id to the frontend RunDashboardManager, which is now capable of checking if the corresponding run configuration is configured to be shown in it.

Q: why did we not use a dedicated rpc inside run dashboard subsystem and decided to change something in debugger instead?
A: because sending the ID through different RPC cll will effectively introduce a race condition out of our control. If debugger will first do its business and then receive an update, that the content must be dispatched to the Services view instead of its own toolwindow, it might be too late. The second reason is a very-very-very high code complexity around the execution module. The current flow can be observed and understood by a single person. Adding even more obscure abstraction will likely increase the maintenance costs even more :(

Q: what about the run toolwindow content?
A: so far it is not displayed in services correctly. There must be the same flow for it, except that both content and descriptor will belong to backend (run toolwindow is not split properly).

GitOrigin-RevId: 9ab1935a9b5f995faee1d1706c46507dccd7dc70