[rdct] IJPL-198085: Respect `CAN_BE_REMOVED_WHEN_COLLAPSED` when merging frontend's and backend's foldings
Since `CAN_BE_REMOVED_WHEN_COLLAPSED` has not been extent to the merged foldings, the duplicated foldings (≈all Java foldings, for example) have lost that property after merge.

That led to the invalid state of the folding model after some backend changes had been applied: if the folding was extended (e.g. with new import) the old folding has not been removed and the new one will be in an expanded state.

That led to the invalid caret movement on the frontend: some lines have been magically skipped.

GitOrigin-RevId: b899c913d73d13beba27fe3f74e7f21d00730349