IJPL-167857 Allow `EditorSelectWord` to be executed as a nested action
Only `FrontendOnly` or `FrontendFirst` actions can be used in nested execution (when one action handler calls another).

By default, `EditorSelectWord` is backend-only, with `FrontendFirst` used only for split languages.
It cannot be `FrontendFirst` by default because extending a selection beyond the current word may require backend information.

However, in the case of a double-click (`EditorSelectWordAtCurrentCaret`), this action can be executed on the frontend.
To ensure correct behavior, we need to allow `EditorSelectWord` to run in nested execution on the frontend.

GitOrigin-RevId: 6ed97e0b21f8c47c81e856d23d6e3b0a2f8f8538