IJPL-207199 Popups are shown on top of other windows
In Wayland, focus transfer between windows is not "atomic": indeterminate amount of time can pass between focus lost and gain, and the corresponding even would have null for the 'opposite window' property. When a popup is shown, the keyboard focus remains on the parent window.
 So if a popup is shown as a top-level window, make sure it is NOT canceled on window deactivation.
The existing cancellation logic remains valid for regular popups, however.

GitOrigin-RevId: fd55354e5b3930736c4b7f3442d91d9f96506bd1