[platform UI] IJPL-207764 Use deferredMask to update composite icon size
If an icon was deferred, but later was evaluated,
then its size might change. This means that sizeIsDirty = false
might be outdated.

Because a deferred icon has no way to tell its containing
icon to update its size, we need to query its status
when we need the size. Because some component
icons may be deferred icons themselves,
we also need to track indirect components recursively.

Use the new utility function to do this.

GitOrigin-RevId: 32d019f3b25002f0bf117e15e822a9b77a25cd91