IJPL-189375 allow checking IndexingFlag during indexing (not only during initial scanning)
There is only one place in production code where allowCheckingForOutdatedIndexesUsingFileModCount changes behavior: WriteOnlyFileIndexingStampImpl.isFileChanged, which has exactly one usage in IndexingFlag.isFileChanged which is used only from FileBasedIndexImpl.getIndexingState. In other words, we have only modified `FileBasedIndexImpl.getIndexingState` with this change, which is now allowed to report that index is not up-to-date if the file has changed in the VFS.

Possible undesired side effect is that FileBasedIndexImpl.getIndexingState now will report "OUTDATED" instead of "NOT_INDEXED" which does not look like a problem at the moment.

GitOrigin-RevId: 8c8e01175742439b2533e973523a2788ae60d53d