IJPL-194226 IJent: avoid failing tests by logging "The process ... suddenly exited with the code ..."
Various coroutine scopes deliberately use the antipattern "control flow exception" (`IjentUnavailableException`) to indicate on terminated connections with IJent. It helped us find, fix and prevent bugs caused by a flaw in the design of Kotlin coroutines that is known as the "silent killer".

The drawback of this approach is that these control flow exceptions appear here and there, especially in cases when such errors are expected, i.e., when the process was killed by the IDE.

An error log message logged during running a test causes the test failure. This commit prevents logging uncaught `IjentUnavailableException` errors.

GitOrigin-RevId: fca1360aab7946592afe7c83328790956ef2ce40