[platform UI, Switcher] IJPL-195117 Fix the Switcher with tab limit = 1
Since the regular Recent Files popup works as intended in this mode,
we can fix the Switcher by making it show the same set of files.

So in the Switcher (unpinned) mode it now requests
RECENTLY_OPENED if the tab limit = 1 and should show
the same files as the Recent Files popup.

GitOrigin-RevId: 5e6108b531026133390279cb592b261209632521