[UI, PV] IJPL-158511 Introduce project.view.toolwindow.split
To make it work, we need the tool window factory in all modes.
So we define it in the shared module, and then delegate to the right
implementation.

If the key is disabled, it's simple: just call the old implementation.
If the key is enabled, we use a frontend-only service to create
the tool window. We use serviceOrNull for this.

GitOrigin-RevId: c32f473144c5e2e4bda5785de43056edf0ddd64f