[plugin model] if a plugin is not loaded, ensure that other modules which depend on its required modules aren't loaded too (IJPL-202345)
Required (and embedded) modules come before the main module in sortedModulesWithDependencies to ensure that we can determine whether the plugin can be loaded or not when processing its main module. However, if the plugin is not loaded for some reason (e.g., because it's not included in PluginInitializationContext.explicitPluginSubsetToLoad), we need to remove its content modules from 'enabledRequiredContentModules' to ensure that modules of other plugins which depend on these modules won't be loaded.

This also fixes problems with loading 'intellij.go.shell' plugins in Fleet tests: Fleet uses 'idea.load.plugins.id' to specify plugins to load, and 'com.jetbrains.sh' wasn't loaded because of that, but its embedded module 'intellij.sh.core' wasn't removed from 'enabledRequiredContentModules', so 'intellij.go.shell' was loaded.

GitOrigin-RevId: be3efb575788c0b3749365b903e6e316f54d4912