[todo] IJPL-208085 Optimize listing TODO items through RPC
Introduced a small data class responsible for collecting the TODO item offsets in file. Instead of one huge read block, now there are several of them - one for accessing document, one for filtering the TODO items, and one for serializing preview of each TODO item.
This is still work in progress and might be a subject of change soon.

GitOrigin-RevId: d93e189268689e36e5e4270f5eebd6f9c03f1e6e