[platform UI, API] IJPL-208412 Fix TreeState sync selection restore
If it's impossible to fully restore the path to be selected,
we must at least try to find a reasonable fallback.
It's either the child of the same parent with the same index
(corresponds to selecting the next node on deletion)
or at least the parent (if all children were deleted).

Implement and add unit tests.

GitOrigin-RevId: b26801f39c4de8be876d93131c5092f7961f2471