[platform UI, API] IJPL-208412 Refactor TreeState: rename and extract
Rename findMatchedPath because from the previous
name it wasn't clear that it's only used for selected paths
and not for expanded paths. Because the logic
for expanded and selected paths is different,
it's important that we won't accidentally call
this function for expanded paths.
One example when it would be bad is when an expanded
path is removed. For the selection, after the fix
we just select the next node. But for the expanding,
it would be very confusing and undesirable:
if the expanded node is deleted, we should just forget it.

GitOrigin-RevId: f98fc7c0c5513a449051cda399a326541002be77