[kotlin] do not perform checksum on tests generation
It is too costly to call it on each run since it
takes much longer than the actual test generation.
For a reference, checksum is ~20s while the test generation ~3s.
The actual artifacts generation happens on the compiler version update,
but this is a separate process that has no common with tests generation

GitOrigin-RevId: ddfc45657dd49d1c31c77bd9f8c8a77038facd5a