[unit tests] IJPL-209375 Convert IconScaleTest to BasePlatformTestCase
This is needed because DeferredIconImpl
requires an Application to access dispatching
services.

Why it worked fine before is not clear,
but the most likely reason is that it used
"leftovers" from previous tests.
Now something has changed in test ordering.

This is also needed because scaling uses a couple
Registry keys.

Because it's JUnit 4 now, we call manageState
manually before/after every test. Thanks to this,
we can get rid of manual cleanup code in tearDown.
That code was added only so that the tests don't
interfere with each other within this class.

GitOrigin-RevId: 56a716d25a7ec0c4c077b261739db25335582de1