[Gradle][WSM] IDEA-377941 fix unloading modules by removing @Parent connection with Gradle project in GradleModuleEntity
Fixes the problem: when a module is unloaded, `EntityTreeUtilKt#createEntityTreeCopy` is called for ModuleEntity. This method fails because: "some elements weren't copied because their additional parents are located in a separate tree: $deferred". In this case, the mentioned element is `GradleModuleEntity`, because it also has another parent - `GradleProjectEntity`. So, I replaced a connection with it by its symbolic ID.

Fixes the test:
* org.jetbrains.plugins.gradle.importing.GradleMiscImportingTest.testUnloadedModuleImport[with Gradle-9.0.0]

Probably, these two tests were also failing because of the same issue:
* com.intellij.workspaceModel.integrationTests.tests.aggregator.gradle.changes.ReloadGradleModulesTest.reloadGradleModules
* com.intellij.workspaceModel.integrationTests.tests.aggregator.gradle.changes.UnloadGradleModulesTest.unloadGradleModules
(unable to check locally on macOS if they are fixed)

Tests for GradleModuleEntity:
* org.jetbrains.plugins.gradle.importing.syncAction.GradleContentRootSyncContributorTest

GitOrigin-RevId: 35acf502a699f1dc9e75d207f200e69cb93aa57c