[JEWEL-1027] Fix the branch comparison script
The script had issues in picking up which commits to look for
and did not account for things like the time difference between
a commit happening on master and its cherry picks happening on
a branch.

This fixes the issue and cleans up the script code, improving
logging too. It also adds a few flags to make the output of the
script a bit more manageable.

closes https://github.com/JetBrains/intellij-community/pull/3252

GitOrigin-RevId: e8318e2cddbd84c22b94d93df4a2985ef04c8de8