[JEWEL-1027] Fix the branch comparison script
The script had issues in picking up which commits to look for
and did not account for things like the time difference between
a commit happening on master and its cherry picks happening on
a branch.

This fixes the issue and cleans up the script code, improving
logging too. It also adds a few flags to make the output of the
script a bit more manageable.