[UI, HTML rendering] IJPL-209887 Clean up adaptive image API
Most of this stuff is just internals.
The only exception is DataUrl which looks useful,
so let's expose it, hiding data class internals.

In withRasterizationIntention, computeIfAbsent had to be replaced
by getOrPut because the former would expose the SAM implementation.
While this looks like a bug either in the compiler or
in ApiCheckTest, it's easier to just work around it,
because there's no concurrencty involved, so getOrPut is OK.

Drop SVGUtil because we already have a similar hack
to work around compiler quirks. Just move the classes
to the package where the hack is, because it's package private.

GitOrigin-RevId: cfbd9da08ad1edacdd2b8fe6a5316b19ad122a2b