Optimize existing module and content entity lookup
There was an upstream PR to do this that got merged after a
couple of refactorings that made it not very straightforward
to pull, so I just recreated the changes. Also slightly
modified the change after tests detected a slight change
in behaviour.

Upstream PR: https://github.com/JetBrains/intellij-community/pull/3026

Test: existing integration tests
Bug: 384022658
Change-Id: I3b37983a93e3f1165aa55a7b7357e81b7130ee7a