[UI, HTML rendering] IJPL-209887 Fix slow ops in source resolution
We can't just convert the origin to the source on the EDT
because of the VFS slow ops involved.

So make it suspend and rely on the origin for comparison.
Here we have to assume that origin equality means source equality.
In most cases it's true, and if it turns out that sometimes it isn't,
we'll have to deal with it in some different way.

GitOrigin-RevId: 9424d39a221c62781cbb69ca193dc09a474f01d2