[terminal] IJPL-211122 Refactor: move TerminalTitle to TerminalView
Initially, having TerminalTitle separated from the TerminalView sounded like a good idea because title is needed only to present the terminal component in the tool window or editor tab. And it is not needed when TerminalView is used in some other place.
But found out that it is much more suitable to have it in the TerminalView.

GitOrigin-RevId: 86da64e202a916028f0a192352e2234415a52351