[gradle][IDEA-379633] (kotlin) Smarter repository merging, highlighting changes before and after quickfix, code cleanup.
Merging repositories will just append to the existing list of repositories. It can add duplicates but will not add lines if there is an exact overlap with existing ones such that the order of both existing and moved repositories is maintained. See test 'testRepositoriesMergeToExistingDependencyResolutionManagementOverlap'.
Highlight added lines after a quickfix, only highlight the 'repositories' part of the method call instead of the whole block.
Adjust tests.

Space-RevId: 000f8b254174d637cde828b0c19e65ce6ef247c3

GitOrigin-RevId: 4270260db4d718ae533aede32375dde74d76079f