AP-7777 Use FUS metadata/dictionary update routine from ap-validation-all library.
This is the first step to migrate the FUS client library into its new component structure. For now, we expose an internal message bus to communicate internal metadata/dictionary update events to IntelliJ. Once all FUS logic is part of the library, these types of events should not be exposed to IntelliJ anymore. The library will be self-contained. See `FusComponentProvider` for component definitions and glue code.

Since it's now part of the library, code that was responsible for tracking last modification timestamps of metadata/dictionaries has been removed.

GitOrigin-RevId: 10e8aa26c619b8639c20d39f8d3a0bb522bcfb93