[jdk][IJPL-197987] sdk quick fix: use the actual SDK name instead of the suggested name sdk
The actual SDK name could be different from the suggested SDK name. When creating a new SDK instance, we use JdkTable to choose the correct name for the SDK that will not conflict with other JDKs on the system. The `getSuggestedSdkName` name does not take into account the current JdkTable. This could lead to a situation where we applied a quick fix and selected a JDK instance that is not represented in the JdkTable because we did not save the JDK in the table due to a jdk name conflict.

GitOrigin-RevId: 75fc9aebf40e50e19a2b53d5911259e5d879bfb3