IJPL-211400 update dynamic code usages text for Symbol API usages search to "Potential usages"
While currently the term "dynamic usages" is used for this concept, it is a bit ambiguous. Dynamic usages could refer to either potential usages, or meta programming usages. I believe the term "Potential usages" is clearer from a UI perspective about the grouping.

This change is only made for the new search targets because there are currently no uses of it, and it will only apply to new uses of the API.

GitOrigin-RevId: e820b5dcf25d1978c168205696263f1b987cf3c4