RUBY-35082 do not duplicate processing of template language elements
`isApplicableOccurrence` is attempting to filter out and occurrences that should not be considered text occurrences. It does this by excluding those that have references or declarations inside them.

When searching for text occurrences in template language files, multiple occurrences can be found for the same text/offset combination. However, we wish to exclude the occurrence if any of them have references.

This approach excludes an occurrence if it is not the "most specific" leaf at the offset. In general this should work, but it is worth noting that it is not the same behaviour that we said we want (from above), nor the same behaviour from `PsiFile#findReferenceAt` (which is how the PSI-based find usage APIs filter text occurrences).

GitOrigin-RevId: 7c6b840c0b2c727cbd4636426c32a2ba8780d0ae