[run dashboard & split mode] IJPL-184406 Avoid creating a secondary configuration node when a backend (from RUN toolwindow) content descriptor is replaced by frontend (from split DEBUGGER) one or vice versa
Since they appear on different sides of the split IDE, the old logic relying on the ability to reuse debugger's descriptor by run content and vice versa is now broken. To prevent spawning a duplicate node in the tree when the same configuration is first run and then debugged, we need to make sure that only same origin descriptors exchange happens according to the old logic. If they are different, we just replace one by the other. That means we pretend the content has been reused. It actually is not (the real contents are in fact created in different processes). But this change restores the original bahviour in a new world where the debugger is split and the run toolwindow is not

GitOrigin-RevId: 51aa9e65ae0f1e71b4fbf4a4acdc83b06371f4d3